refactor: extract get page api
