+++ frontend/package.json
@@ -8,15 +8,16 @@
     "@testing-library/user-event": "^12.1.10",
     "axios": "^0.24.0",
     "create-react-app": "^4.0.3",
+    "npm": "^8.5.3",
     "qs": "^6.10.2",
     "react": "^17.0.2",
     "react-cookie": "^4.1.1",
     "react-dom": "^17.0.2",
     "react-markdown": "^7.1.2",
+    "react-redux": "^7.2.6",
     "react-router": "^6.2.1",
     "react-router-dom": "^6.2.1",
     "react-scripts": "4.0.3",
-    "react-redux": "^7.2.6",
     "redux": "^4.1.2",
     "redux-thunk": "^2.4.1",
     "web-vitals": "^1.0.1"
