+++ frontend/src/blog/api/getPage.js
@@ -1,0 +1,17 @@
+import qs from "qs";
+import axios from "axios";
+
+export const getPage = ({url, page}) => {
+
+    const queryString = qs.stringify(
+        {
+            page: page,
+            size: 10
+        }
+    )
+    console.log(page);
+
+    return axios.get(url + queryString)
+        .then(response => response)
+        .catch(reason => console.log(reason));
+}
