+++ frontend/src/blog/components/Login.js
@@ -1,7 +1,7 @@
 import React from "react";
 import axios from "axios";
 import {useState} from "react";
-import {LOGIN} from "../support/UrlUtils";
+import {LOGIN} from "../../support/UrlUtils";
 
 function Login(props) {
