+++ frontend/src/blog/components/Register.js
@@ -1,7 +1,7 @@
 import React from 'react'
 import axios from 'axios';
 import { useState } from 'react';
-import { REGISTER } from '../support/UrlUtils';
+import { REGISTER } from '../../support/UrlUtils';
 
 
 function Register() {
